feat: added SWR and implemented it
