+++ package.json
@@ -22,6 +22,7 @@
     "react-loading-skeleton": "^3.2.0",
     "regression": "^2.0.1",
     "sweetalert": "^2.1.2",
+    "swr": "^2.2.4",
     "typescript": "4.9.5",
     "zod": "^3.21.4"
   },
