+++ src/components/ReportSection/ReportTable/index.tsx
@@ -1,80 +1,115 @@
 "use client";
 
-import { BASE_URL } from 'src/services/api';
-import useSession from '@/hooks/useSession';
-import { useTheme } from 'next-themes';
 import { useState } from 'react';
 import Report from '@/types/Report';
-import ReportTable from '..';
-import swal from 'sweetalert';
-
+import { FaArrowUp, FaArrowDown } from 'react-icons/fa';
 
 type TableProps = {
     reports: Report[];
+    selectedReports: number[];
+    setSelectedReports: React.Dispatch<React.SetStateAction<number[]>>;
+    sortBy: { key: string; desc: boolean };
+    setSortBy: React.Dispatch<React.SetStateAction<{
+        key: string;
+        desc: boolean;
+    }>>;
 };
 
 
-const ReportSection: React.FC<TableProps> = ({ reports }) => {
-    const { theme } = useTheme();
-    const session = useSession();
+const ReportTable: React.FC<TableProps> = ({ reports, selectedReports, setSelectedReports, sortBy, setSortBy }) => {
 
-    const [selectedReports, setSelectedReports] = useState<number[]>([]);
+    const collumns = ['ID', 'Autor', 'Descrição', 'ID Questão', 'Criado em', 'Resolvido'];
 
-    const handleMarkAsResolve = async () => {
-
-        const res = await fetch(BASE_URL + '/question-reports/review', {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-                Authorization: `Bearer ${session.token}`
-            },
-            body: JSON.stringify({
-                question_ids: selectedReports,
-            })
-        });
-
-        if (res.status === 200)
-            swal({
-                title: 'Resolvido!',
-                text: `${selectedReports.length > 1 ? 'Os reports foram marcados como resolvidos' : 'O report foi marcado como resolvido'}.`,
-                icon: 'success',
-                className: theme === 'dark' ? 'swal-dark' : '',
-                timer: 2000
-            });
-        else
-            swal({
-                title: 'Erro!',
-                text: 'Algo correu mal ao marcar o(s) teu(s) report(s). Por favor, tenta novamente.',
-                icon: 'error',
-                className: theme === 'dark' ? 'swal-dark' : ''
-            });
-        console.log(selectedReports);
+    // select a single row
+    const toggleRow = (id: number) => {
+        setSelectedReports((prevselectedReports) =>
+            prevselectedReports.includes(id)
+                ? prevselectedReports.filter((rowId) => rowId !== id)
+                : [...prevselectedReports, id]
+        );
     };
 
-    const handleDelete = async () => {
-        if (selectedReports.length === 0) return;
+    // select all rows
+    const handleSelectAll = () => {
+        const allRowIds = reports.map((row) => row.id);
+        setSelectedReports((prevselectedReports) =>
+            prevselectedReports.length === allRowIds.length ? [] : allRowIds
+        );
+    };
+
+    // sort table
+    const handleSort = (collumn: string) => {
+        let desc = false;
+        if (sortBy.key === collumn && sortBy.desc) {
+            desc = true;
+        }
+        setSortBy({ key: collumn, desc });
     };
 
     return (
-        <>
-            <div className="flex gap-x-2">
-                <button
-                    className="bg-gray-400 hover:bg-gray-500 text-white px-4 py-2 rounded-md shadow-md"
-                    onClick={handleMarkAsResolve}
-                >
-                    Resolver
-                </button>
-                <button className="bg-red-500 hover:bg-red-500 text-white px-4 py-2 rounded-md shadow-md"
-                    onClick={handleDelete}
-                >
-                    Apagar
-                </button>
-            </div>
-            <div className="flex flex-col w-3/4">
-                <ReportTable reports={reports} selectedReports={selectedReports} setSelectedReports={setSelectedReports} />
-            </div>
-        </>
+        <table className="min-w-full border border-gray-300">
+            {reports.length === 0 ? (
+                <thead>
+                    <tr>
+                        <th className="px-4 py-2">Nenhum report encontrado</th>
+                    </tr>
+                </thead>
+            ) : (
+                <>
+                    <thead>
+                        <tr>
+                            <th className="px-4 py-2">
+                                <input
+                                    type="checkbox"
+                                    checked={selectedReports.length === reports.length}
+                                    onChange={handleSelectAll}
+                                />
+                            </th>
+                            {collumns.map((collumn: string) => (
+                                <th
+                                    className="px-4 py-2 cursor-pointer"
+                                    onClick={() => handleSort(collumn)}
+                                    key={collumn}
+                                >
+                                    {collumn}
+                                    {sortBy.key === collumn &&
+                                        (!sortBy.desc ? (
+                                            <FaArrowUp className="inline ml-1" />
+                                        ) : (
+                                            <FaArrowDown className="inline ml-1" />
+                                        ))}
+                                </th>
+                            ))}
+                        </tr>
+                    </thead>
+                    <tbody>
+                        {reports.map((report: Report) => (
+                            <tr
+                                key={report.id}
+                                className={selectedReports.includes(report.id) ? 'bg-blue-200' : ''}
+                            >
+                                <td className="border px-4 py-2">
+                                    <input
+                                        type="checkbox"
+                                        checked={selectedReports.includes(report.id)}
+                                        onChange={() => toggleRow(report.id)}
+                                    />
+                                </td>
+                                <td className="border px-4 py-2">{report.id}</td>
+                                <td className="border px-4 py-2">{report.user}</td>
+                                <td className="border px-4 py-2">{report.reason}</td>
+                                <td className="border px-4 py-2">{report.question.id}</td>
+                                <td className="border px-4 py-2">{report.created_at}</td>
+                                <td className="border px-4 py-2">{report.solved == 1 ? 'Sim' : 'Não'}</td>
+
+                            </tr>
+                        ))}
+                    </tbody>
+                </>
+            )}
+
+        </table>
     );
 };
 
-export default ReportSection;
+export default ReportTable;
